[FIX] Fixing Mason for Windows.
git-svn-id: http://svn.seqan.de/seqan/trunk@14323 e6417c60-b987-48fd-844e-b20f0fcc1017

+++ extras/apps/mason2/fragment_generation.h
@@ -273,7 +273,7 @@ void NormalFragmentSamplerImpl::_generate(Fragment & frag, int rId, unsigned con
     unsigned const MAX_TRIES = 1000;
     for (unsigned tryNo = 0; tryNo < MAX_TRIES; ++tryNo)
     {
-        fragLength = pickRandomNumber(rng, pdf);
+        fragLength = static_cast<int>(pickRandomNumber(rng, pdf));
         if (fragLength <= 0 || fragLength > (int)contigLength)
             continue;  // Try again
